# docs/swagger.yaml
definitions:
...
  github.com_VinCPR_backend_api.academicYearResponse:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
    type: object
  github.com_VinCPR_backend_api.attendingResponse:
    properties:
      attending_id:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      mobile:
        type: string
    type: object
...
  github.com_VinCPR_backend_api.blockResponse:
    properties:
      academic_year_id:
        type: integer
      created_at:
        type: string
      name:
        type: string
      period:
        type: integer
    type: object
...
  github.com_VinCPR_backend_api.createAcademicYearRequest:
    properties:
      end_date:
        type: string
      name:
        type: string
      start_date:
        type: string
    required:
    - end_date
    - name
    - start_date
    type: object
  github.com_VinCPR_backend_api.createAttendingRequest:
    properties:
      attending_id:
        type: string
      email:
        type: string
      firstname:
        type: string
      lastname:
        type: string
      mobile:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - attending_id
    - email
    - firstname
    - lastname
    - mobile
    - password
    type: object
  github.com_VinCPR_backend_api.createGroupRequest:
    properties:
      academic_year_name:
        type: string
      name:
        type: string
    required:
    - academic_year_name
    - name
    type: object
  github.com_VinCPR_backend_api.createGroupToBlockRequest:
    properties:
      academic_year_name:
        type: string
      block_name:
        type: string
      group_name:
        type: string
      period_name:
        type: string
    type: object
...
  github.com_VinCPR_backend_api.getAcademicCalendarResponse:
    properties:
      academic_calendar_events:
        items:
          $ref: '#/definitions/github.com_VinCPR_backend_api.academicCalendarEventResponse'
        type: array
      academic_year:
        $ref: '#/definitions/github.com_VinCPR_backend_api.academicYearResponse'
    type: object
  github.com_VinCPR_backend_api.groupResponse:
    properties:
      academic_year_name:
        type: string
      created_at:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /academic_year:
    post:
      consumes:
      - application/json
      description: create new academic year
      parameters:
      - description: 'input required: academic year name, start date, end date'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createAcademicYearRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.academicYearResponse'
      summary: create new academic year
      tags:
      - AcademicCalendar
  /academic_year/calendar:
    get:
      consumes:
      - application/json
      description: return the list of events of an academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.getAcademicCalendarResponse'
            type: array
      summary: return the list of events of an academic year
      tags:
      - AcademicCalendar
  /academic_year/list:
    get:
      consumes:
      - application/json
      description: list created academic year
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.academicYearResponse'
            type: array
      summary: list created academic year
      tags:
      - AcademicCalendar
  /attending:
    post:
      consumes:
      - application/json
      description: create new Attending
      parameters:
      - description: 'input required: attending user_id, firstname, lastname, mobile'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createAttendingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.attendingResponse'
      summary: create new Attending
      tags:
      - Attendings
  /attending/info:
    get:
      consumes:
      - application/json
      description: provide detail of an attending given email address
      parameters:
      - description: attending email address
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.attendingResponse'
      summary: provide detail of an attending given email address
      tags:
      - Attendings
  /attending/list/attending_id:
    get:
      consumes:
      - application/json
      description: list created attending order by attending id
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.attendingResponse'
            type: array
      summary: list created attending order by attending id
      tags:
      - Attendings
  /attending/list/name:
    get:
      consumes:
      - application/json
      description: list created attending
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.attendingResponse'
            type: array
      summary: list created attending
      tags:
      - Attendings
  /block/list:
    get:
      consumes:
      - application/json
      description: list blocks in that academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.blockResponse'
            type: array
      summary: list blocks in that academic year
      tags:
      - Blocks
  /group:
    post:
      consumes:
      - application/json
      description: create new group
      parameters:
      - description: 'input required: academic year name, group name'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.groupResponse'
      summary: create new group
      tags:
      - Groups
  /group/list:
    get:
      consumes:
      - application/json
      description: list groups in that academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.groupResponse'
            type: array
      summary: list groups in that academic year
      tags:
      - Groups
  /group_to_block:
    post:
      consumes:
      - application/json
      description: create new group to block
      parameters:
      - description: 'input required: academic year name, groupid, blockid'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createGroupToBlockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
      summary: create new group to block
      tags:
      - GroupToBlock
  /group_to_block/list/academic_year:
    get:
      consumes:
      - application/json
      description: list groups and their blocks in that academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.groupToBlockResponse'
            type: array
      summary: list groups and their blocks in that academic year
      tags:
      - GroupToBlock
  /group_to_block/list/block:
    get:
      consumes:
      - application/json
      description: list groups in that block
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: period name
        in: query
        name: periodName
        required: true
        type: string
      - description: block name
        in: query
        name: blockName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.groupToBlockResponse'
            type: array
      summary: list groups in that block
      tags:
      - GroupToBlock
  /group_to_block/list/group:
    get:
      consumes:
      - application/json
      description: list group_to_block of that block
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: group name
        in: query
        name: groupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.groupToBlockResponse'
            type: array
      summary: list group_to_block of that block
      tags:
      - GroupToBlock
  /hospital:
    post:
      consumes:
      - application/json
      description: create new hospital
      parameters:
      - description: 'input required: hospital name, description, address'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createHospitalRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.hospitalResponse'
      summary: create new hospital
      tags:
      - Hospitals
  /hospital/list:
    get:
      consumes:
      - application/json
      description: list created hospital
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.hospitalResponse'
            type: array
      summary: list created hospital
      tags:
      - Hospitals
  /period/list:
    get:
      consumes:
      - application/json
      description: list periods in that academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.periodResponse'
            type: array
      summary: list periods in that academic year
      tags:
      - Periods
  /rotation/attending:
    get:
      consumes:
      - application/json
      description: list calendar events for an attending in an academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: attending ID
        in: query
        name: attendingID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list calendar events for an attending in an academic year
      tags:
      - ClinicalRotation
  /rotation/design:
    post:
      consumes:
      - application/json
      description: provide templates of each period to auto generate calendar for
        all students
      parameters:
      - description: 'input required: academic year name, templates for each period'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.designRotationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
      summary: design clinical rotation
      tags:
      - ClinicalRotation
  /rotation/detail:
    get:
      consumes:
      - application/json
      description: provide students and attending info and the clinical info
      parameters:
      - description: clinical rotation event ID
        in: query
        name: eventID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventDetailResponse'
      summary: provide detail of a clinical rotation event
      tags:
      - ClinicalRotation
  /rotation/list/day:
    get:
      consumes:
      - application/json
      description: list clinical rotation event in that day
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: day in format YYYY-MM-DD
        in: query
        name: day
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list clinical rotation event in that day
      tags:
      - ClinicalRotation
  /rotation/reset:
    post:
      consumes:
      - application/json
      description: reset all rotation events of an academic year
      parameters:
      - description: 'input required: academic year name'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.resetRotationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
      summary: reset clinical rotation
      tags:
      - ClinicalRotation
  /rotation/student:
    get:
      consumes:
      - application/json
      description: list calendar events for a student in an academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: student ID
        in: query
        name: studentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.clinicalRotationEventResponse'
            type: array
      summary: list calendar events for a student in an academic year
      tags:
      - ClinicalRotation
  /service:
    post:
      consumes:
      - application/json
      description: create new service
      parameters:
      - description: 'input required: service hospitalID, specialtyID, name, description'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createServiceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.serviceResponse'
      summary: create new service
      tags:
      - Services
  /service/get:
    get:
      consumes:
      - application/json
      description: get created service of hospital and specialty
      parameters:
      - description: specialty
        in: query
        name: specialty
        required: true
        type: string
      - description: hospital
        in: query
        name: hospital
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceResponse'
            type: array
      summary: get created service of hospital and specialty
      tags:
      - Services
...
  /service/list/hospital:
    get:
      consumes:
      - application/json
      description: list created service
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceResponse'
            type: array
      summary: list created service
      tags:
      - Services
  /service/list/specialty:
    get:
      consumes:
      - application/json
      description: list created service
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceResponse'
            type: array
      summary: list created service
      tags:
      - Services
  /service/list/specialty_and_hospital:
    get:
      consumes:
      - application/json
      description: list created service
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceResponse'
            type: array
      summary: list created service
      tags:
      - Services
  /service_to_attending:
    post:
      consumes:
      - application/json
      description: create new service to attending
      parameters:
      - description: 'input required: serviceID, attendingID'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createServiceToAttendingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.serviceToAttendingResponse'
      summary: create new service to attending
      tags:
      - ServiceToAttendings
  /service_to_attending/list/all:
    get:
      consumes:
      - application/json
      description: list created services to attendings
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceToAttendingResponse'
            type: array
      summary: list created services to attendings
      tags:
      - ServicesToAttendings
  /service_to_attending/list/attending_id:
    get:
      consumes:
      - application/json
      description: list created services to attendings
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceToAttendingResponse'
            type: array
      summary: list created services to attendings
      tags:
      - ServicesToAttendings
  /service_to_attending/list/service_id:
    get:
      consumes:
      - application/json
      description: list created services to attendings
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.serviceToAttendingResponse'
            type: array
      summary: list created services to attendings
      tags:
      - ServicesToAttendings
  /specialty:
    post:
      consumes:
      - application/json
      description: create new specialty
      parameters:
      - description: 'input required: specialty name, description, address'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createSpecialtyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.specialtyResponse'
      summary: create new specialty
      tags:
      - Specialties
  /specialty/list:
    get:
      consumes:
      - application/json
      description: list created specialty
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.specialtyResponse'
            type: array
      summary: list created specialty
      tags:
      - Specialties
  /student:
    post:
      consumes:
      - application/json
      description: create new student
      parameters:
      - description: 'input required: student user_id, student_id, firstname, lastname,
          mobile'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createStudentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.studentResponse'
      summary: create new student
      tags:
      - Students
  /student/info:
    get:
      consumes:
      - application/json
      description: provide detail of a student given email address
      parameters:
      - description: student email address
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.studentResponse'
      summary: provide detail of a student given email address
      tags:
      - Students
  /student/list/name:
    get:
      consumes:
      - application/json
      description: list created student
      parameters:
      - description: page number
        in: query
        name: pageNumber
        required: true
        type: string
      - description: page size
        in: query
        name: pageSize
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.studentResponse'
            type: array
      summary: list created student
      tags:
      - Students
...
  /student_to_group:
    post:
      consumes:
      - application/json
      description: create new student to group
      parameters:
      - description: 'input required: academic year name, studentid, groupid'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createStudentToGroupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.studentToGroupResponse'
      summary: create new student to group
      tags:
      - StudentToGroup
  /student_to_group/list/academic_year:
    get:
      consumes:
      - application/json
      description: list students and their groups in that academic year
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.studentToGroupResponse'
            type: array
      summary: list students and their groups in that academic year
      tags:
      - StudentToGroup
  /student_to_group/list/group:
    get:
      consumes:
      - application/json
      description: list students of that group
      parameters:
      - description: academic year name
        in: query
        name: academicYearName
        required: true
        type: string
      - description: group name
        in: query
        name: groupName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.studentToGroupResponse'
            type: array
      summary: list students of that group
      tags:
      - StudentToGroup
  /student_to_group/list/student:
    get:
      consumes:
      - application/json
      description: list groups of that student
      parameters:
      - description: studentID
        in: query
        name: studentID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github.com_VinCPR_backend_api.studentToGroupResponse'
            type: array
      summary: list groups of that student
      tags:
      - StudentToGroup
  /users:
    post:
      consumes:
      - application/json
      description: upsert wallet address
      parameters:
      - description: create new user request requires username, password, and role
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.createUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.userResponse'
      summary: upsert new user
      tags:
      - Authentication
  /users/login:
    post:
      consumes:
      - application/json
      description: upsert wallet address
      parameters:
      - description: login request requires username, password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github.com_VinCPR_backend_api.loginUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github.com_VinCPR_backend_api.loginUserResponse'
      summary: upsert new user
      tags:
      - Authentication
swagger: "2.0"